Moved reducers to own dir
