+++ src/reducers/CurrentBanListReducer.ts
